refactor: fix the code to match the feedback
