+++ src/main/kotlin/lotto/domain/model/Lotto.kt
@@ -9,7 +9,21 @@ class Lotto {
         FOUR_MATCH(4, 50_000),
         FIVE_MATCH(5, 1_500_000),
         FIVE_MATCH_PLUS_BONUS(5, 30_000_000),
-        SIX_MATCH(6, 2_000_000_000)
+        SIX_MATCH(6, 2_000_000_000);
+
+        companion object {
+            fun from(selectedBalls: SelectedBalls, lotto: Lotto): Prize {
+                val hasBonusBall = lotto.numbers.contains(selectedBalls.bonus)
+                val matchCount = lotto.numbers.filter { it in selectedBalls.winningNumbers.balls }.size
+                return when (matchCount) {
+                    Lotto.Prize.SIX_MATCH.matches -> Lotto.Prize.SIX_MATCH
+                    Lotto.Prize.FIVE_MATCH.matches -> if (hasBonusBall) Lotto.Prize.FIVE_MATCH_PLUS_BONUS else Lotto.Prize.FIVE_MATCH
+                    Lotto.Prize.FOUR_MATCH.matches -> Lotto.Prize.FOUR_MATCH
+                    Lotto.Prize.THREE_MATCH.matches -> Lotto.Prize.THREE_MATCH
+                    else -> Lotto.Prize.NOTHING
+                }
+            }
+        }
     }
 
     companion object {
